fixed bug: crash on adding var with one compartment
so.  Compartment IO that uses multiple compartments and filters the the original compartment out of the choices for the other IO.  When one compartment this filters to a null.  The null was being passed to a filter to species but causing a break because it didn't exist.  Added is.null catch